honksquad: refactor: extract FontFamilyInfo to its own file
Moves FontFamilyInfo class from the bottom of FontCustomizationManager.cs
into a dedicated file, following one-class-per-file convention.

Closes #326
